refactor: move openai message converting functions to utils
