+++ libs/langchain/langchain/adapters/openai.py
@@ -6,8 +6,6 @@ from typing import (
     AsyncIterator,
     Dict,
     Iterable,
-    List,
-    Mapping,
     Sequence,
     Union,
     overload,
@@ -15,17 +13,11 @@ from typing import (
 
 from typing_extensions import Literal
 
-from langchain.schema.chat import ChatSession
 from langchain.schema.messages import (
-    AIMessage,
     AIMessageChunk,
-    BaseMessage,
     BaseMessageChunk,
-    ChatMessage,
-    FunctionMessage,
-    HumanMessage,
-    SystemMessage,
 )
+from langchain.utils.openai import convert_message_to_dict, convert_openai_messages
 
 
 async def aenumerate(
@@ -36,67 +28,6 @@ async def aenumerate(
     async for x in iterable:
         yield i, x
         i += 1
-
-
-def convert_dict_to_message(_dict: Mapping[str, Any]) -> BaseMessage:
-    role = _dict["role"]
-    if role == "user":
-        return HumanMessage(content=_dict["content"])
-    elif role == "assistant":
-        # Fix for azure
-        # Also OpenAI returns None for tool invocations
-        content = _dict.get("content", "") or ""
-        if _dict.get("function_call"):
-            additional_kwargs = {"function_call": dict(_dict["function_call"])}
-        else:
-            additional_kwargs = {}
-        return AIMessage(content=content, additional_kwargs=additional_kwargs)
-    elif role == "system":
-        return SystemMessage(content=_dict["content"])
-    elif role == "function":
-        return FunctionMessage(content=_dict["content"], name=_dict["name"])
-    else:
-        return ChatMessage(content=_dict["content"], role=role)
-
-
-def convert_message_to_dict(message: BaseMessage) -> dict:
-    message_dict: Dict[str, Any]
-    if isinstance(message, ChatMessage):
-        message_dict = {"role": message.role, "content": message.content}
-    elif isinstance(message, HumanMessage):
-        message_dict = {"role": "user", "content": message.content}
-    elif isinstance(message, AIMessage):
-        message_dict = {"role": "assistant", "content": message.content}
-        if "function_call" in message.additional_kwargs:
-            message_dict["function_call"] = message.additional_kwargs["function_call"]
-            # If function call only, content is None not empty string
-            if message_dict["content"] == "":
-                message_dict["content"] = None
-    elif isinstance(message, SystemMessage):
-        message_dict = {"role": "system", "content": message.content}
-    elif isinstance(message, FunctionMessage):
-        message_dict = {
-            "role": "function",
-            "content": message.content,
-            "name": message.name,
-        }
-    else:
-        raise TypeError(f"Got unknown type {message}")
-    if "name" in message.additional_kwargs:
-        message_dict["name"] = message.additional_kwargs["name"]
-    return message_dict
-
-
-def convert_openai_messages(messages: Sequence[Dict[str, Any]]) -> List[BaseMessage]:
-    """Convert dictionaries representing OpenAI messages to LangChain format.
-
-    Args:
-        messages: List of dictionaries representing OpenAI messages
-
-    Returns:
-        List of LangChain BaseMessage objects.
-    """
-    return [convert_dict_to_message(m) for m in messages]
 
 
 def _convert_message_chunk_to_delta(chunk: BaseMessageChunk, i: int) -> Dict[str, Any]:
@@ -207,19 +138,3 @@ class ChatCompletion:
                 _convert_message_chunk_to_delta(c, i)
                 async for i, c in aenumerate(model_config.astream(converted_messages))
             )
-
-
-def _has_assistant_message(session: ChatSession) -> bool:
-    """Check if chat session has an assistant message."""
-    return any([isinstance(m, AIMessage) for m in session["messages"]])
-
-
-def convert_messages_for_finetuning(
-    sessions: Iterable[ChatSession],
-) -> List[List[dict]]:
-    """Convert messages to a list of lists of dictionaries for fine-tuning."""
-    return [
-        [convert_message_to_dict(s) for s in session["messages"]]
-        for session in sessions
-        if _has_assistant_message(session)
-    ]
