+++ libs/langchain/langchain/chat_models/everlyai.py
@@ -5,7 +5,6 @@ import logging
 import sys
 from typing import TYPE_CHECKING, Dict, Optional, Set
 
-from langchain.adapters.openai import convert_message_to_dict
 from langchain.chat_models.openai import (
     ChatOpenAI,
     _import_tiktoken,
@@ -13,6 +12,7 @@ from langchain.chat_models.openai import (
 from langchain.pydantic_v1 import Field, root_validator
 from langchain.schema.messages import BaseMessage
 from langchain.utils import get_from_dict_or_env
+from langchain.utils.openai import convert_message_to_dict
 
 if TYPE_CHECKING:
     import tiktoken
